requirements.txt
google-cloud-storage
#dlt[clickhouse]
dlt[duckdb]
pandas
dbt-core
dbt-bigquery
kafka-python
#urllib3